add item publish

+++ megumin-main/src/main/java/cn/sabercon/main/domain/entity/Comment.java
@@ -29,7 +29,7 @@ public class Comment extends BaseEntity {
     /**
      * 楼层数
      */
-    private Integer sn;
+    private Long sn;
 
     private String text;
