+++ megumin-main/src/main/java/cn/sabercon/main/domain/model/CommentModel.java
@@ -23,7 +23,7 @@ public class CommentModel {
 
     private Long postId;
 
-    private Integer sn;
+    private Long sn;
 
     private String text;
